+++ src/Utility/BackupExport.php
@@ -25,14 +25,21 @@ namespace MysqlBackup\Utility;
 
 use Cake\Core\Configure;
 use Cake\Datasource\ConnectionManager;
-use Cake\Filesystem\Folder;
 use Cake\Network\Exception\InternalErrorException;
+use MysqlBackup\BackupTrait;
 
 /**
  * Utility to export databases
  */
 class BackupExport
 {
+    use BackupTrait;
+
+    /**
+     * @var \MysqlBackup\Utility\BackupManager
+     */
+    public $BackupManager;
+
     /**
      * Compression type
      * @var bool|string|null
@@ -44,6 +51,12 @@ class BackupExport
      * @var array
      */
     protected $connection;
+
+    /**
+     * Recipient of the email, if you want to send the backup via mail
+     * @var bool|string
+     */
+    protected $emailRecipient = false;
 
     /**
      * Executable command
@@ -77,6 +90,7 @@ class BackupExport
     public function __construct()
     {
         $this->connection = ConnectionManager::getConfig(Configure::read(MYSQL_BACKUP . '.connection'));
+        $this->BackupManager = new BackupManager;
     }
 
     /**
@@ -146,7 +160,7 @@ class BackupExport
         }
 
         $this->compression = $compression;
-        $this->extension = extensionFromCompression($compression);
+        $this->extension = $this->getExtension($compression);
 
         return $this;
     }
@@ -171,17 +185,15 @@ class BackupExport
             '{$DATABASE}',
             '{$DATETIME}',
             '{$HOSTNAME}',
-            '{$TIMESTAMP}'
+            '{$TIMESTAMP}',
         ], [
             $this->connection['database'],
             date('YmdHis'),
             $this->connection['host'],
-            time()
+            time(),
         ], $filename);
 
-        if (!Folder::isAbsolute($filename)) {
-            $filename = Configure::read(MYSQL_BACKUP . '.target') . DS . $filename;
-        }
+        $filename = $this->getAbsolutePath($filename);
 
         if (!is_writable(dirname($filename))) {
             throw new InternalErrorException(__d('mysql_backup', 'File or directory `{0}` not writable', dirname($filename)));
@@ -192,12 +204,12 @@ class BackupExport
         }
 
         //Checks for extension
-        if (!extensionFromFile($filename)) {
+        if (!$this->getExtension($filename)) {
             throw new InternalErrorException(__d('mysql_backup', 'Invalid file extension'));
         }
 
         //Sets the compression
-        $this->compression(compressionFromFile($filename));
+        $this->compression($this->getCompression($filename));
 
         $this->filename = $filename;
 
@@ -220,13 +232,29 @@ class BackupExport
     }
 
     /**
+     * Sets the recipient's email address to send the backup file via mail
+     * @param bool|string $recipient Recipient's email address or `false` to disable
+     * @return \MysqlBackup\Utility\BackupExport
+     * @since 1.1.0
+     * @uses $emailRecipient
+     */
+    public function send($recipient = false)
+    {
+        $this->emailRecipient = $recipient;
+
+        return $this;
+    }
+
+    /**
      * Exports the database
      * @return string Filename path
      * @see https://github.com/mirko-pagliai/cakephp-mysql-backup/wiki/How-to-use-the-BackupExport-utility#export
      * @uses MysqlBackup\Utility\BackupManager::rotate()
+     * @uses MysqlBackup\Utility\BackupManager::send()
      * @uses _getExecutable()
      * @uses _storeAuth()
      * @uses filename()
+     * @uses $BackupManager;
      * @uses $compression
      * @uses $connection
      * @uses $extension
@@ -255,9 +283,13 @@ class BackupExport
 
         chmod($filename, Configure::read(MYSQL_BACKUP . '.chmod'));
 
+        if ($this->emailRecipient) {
+            $this->BackupManager->send($filename, $this->emailRecipient);
+        }
+
         //Rotates backups
         if ($this->rotate) {
-            BackupManager::rotate($this->rotate);
+            $this->BackupManager->rotate($this->rotate);
         }
 
         return $filename;